Modular process orchestrator, code review fixes, cleanup stale files
- Refactor solvers/condensation.py (769→430 lines): extract _condensation_step_core,
  _combined_step_core, _full_step_core, _run_scan helpers; add make_step() composable API
- Fix diffrax.py: strip TomasState bloat from ODE state (use (Nk,Mk) tuple),
  replace try/except tracing trap with explicit n_substeps parameter
- Fix euler.py: mask empty bins in adaptive dt to prevent death spiral
- Fix properties.py: safe division (clamp Nk before jnp.where), remove dead Kn2 variable
- Fix use_tfl type in benchmark scripts (jnp.asarray → Python float for static_argnums)
- Add docs/future_features.md (7 planned improvements)
- Remove 10 stale files: diffrax_original.py, config.py, plotting.py,
  production_driver.py, main.py, diagnose_*.py, run_*_sensitivity.py
- Update CLAUDE.md, architecture.md, PROGRESS.md

+++ benchmarks/python/time_single_scenario.py
@@ -155,7 +155,7 @@ def main():
     enable_org = jnp.float64(1.0)
     enable_inorg = jnp.float64(1.0)
     fn_scale = jnp.float64(1.0)
-    use_tfl = jnp.float64(1.0)
+    use_tfl = 1.0
 
     results = {}
